# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type CardArtworkRequestArtworkStyle = typeof CardArtworkRequestArtworkStyle[keyof typeof CardArtworkRequestArtworkStyle];


export const CardArtworkRequestArtworkStyle = {
  watercolor: 'watercolor',
  'botanical-linework': 'botanical-linework',
  'pressed-petals': 'pressed-petals',
} as const;

export type CardArtworkRequestDecoration = typeof CardArtworkRequestDecoration[keyof typeof CardArtworkRequestDecoration];


export const CardArtworkRequestDecoration = {
  'rose-sprig': 'rose-sprig',
  wildflower: 'wildflower',
  'olive-branch': 'olive-branch',
  none: 'none',
} as const;

export type CardArtworkRequestPaperTone = typeof CardArtworkRequestPaperTone[keyof typeof CardArtworkRequestPaperTone];


export const CardArtworkRequestPaperTone = {
  'warm-cream': 'warm-cream',
  ivory: 'ivory',
  'blush-ivory': 'blush-ivory',
} as const;

export interface CardArtworkRequest {
  artworkStyle: CardArtworkRequestArtworkStyle;
  decoration: CardArtworkRequestDecoration;
  paperTone: CardArtworkRequestPaperTone;
}

export type CardArtworkProvider = typeof CardArtworkProvider[keyof typeof CardArtworkProvider];


export const CardArtworkProvider = {
  gemini: 'gemini',
  'groq-svg': 'groq-svg',
  none: 'none',
} as const;

export interface CardArtwork {
  b64_json: string;
  mimeType: string;
  provider: CardArtworkProvider;
}

export interface CardArtworkError {
  error: string;
}