# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: cards
    description: Text-free card artwork generation
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /card-artwork/generate:
    post:
      operationId: generateCardArtwork
      tags: [cards]
      summary: Generate a text-free decorative card artwork layer
      description: Generates an illustration layer only. Card messages are rendered separately by the client.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CardArtworkRequest"
      responses:
        "200":
          description: Generated artwork image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardArtwork"
        "400":
          description: Invalid artwork request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardArtworkError"
        "502":
          description: Artwork provider failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CardArtworkError"
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    CardArtworkRequest:
      type: object
      properties:
        artworkStyle:
          type: string
          enum: [watercolor, botanical-linework, pressed-petals]
        decoration:
          type: string
          enum: [rose-sprig, wildflower, olive-branch, none]
        paperTone:
          type: string
          enum: [warm-cream, ivory, blush-ivory]
      required:
        - artworkStyle
        - decoration
        - paperTone
    CardArtwork:
      type: object
      properties:
        b64_json:
          type: string
        mimeType:
          type: string
        provider:
          type: string
          enum: [gemini, groq-svg, none]
      required:
        - b64_json
        - mimeType
        - provider
    CardArtworkError:
      type: object
      properties:
        error:
          type: string
      required:
        - error