# lib/api-zod/src/generated/types/cardArtwork.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CardArtworkProvider } from './cardArtworkProvider';

export interface CardArtwork {
  b64_json: string;
  mimeType: string;
  provider: CardArtworkProvider;
}