# lib/api-zod/src/generated/types/cardArtworkRequest.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CardArtworkRequestArtworkStyle } from './cardArtworkRequestArtworkStyle';
import type { CardArtworkRequestDecoration } from './cardArtworkRequestDecoration';
import type { CardArtworkRequestPaperTone } from './cardArtworkRequestPaperTone';

export interface CardArtworkRequest {
  artworkStyle: CardArtworkRequestArtworkStyle;
  decoration: CardArtworkRequestDecoration;
  paperTone: CardArtworkRequestPaperTone;
}